requirements.txt
# AI Companion — Production Dependencies
# ═══════════════════════════════════════════════════════════

# ── Core API ──────────────────────────────────────────────
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
gunicorn>=22.0.0
sse-starlette>=2.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# ── LLM Client ───────────────────────────────────────────
openai>=1.30.0

# ── Cache & Session (Upstash Redis) ──────────────────────
redis>=5.0.0

# ── Database (Neon PostgreSQL — Phase 2) ─────────────────
sqlalchemy>=2.0.0
asyncpg>=0.29.0
# alembic>=1.13.0       # uncomment for migrations

# ── HTTP client ──────────────────────────────────────────
httpx>=0.27.0

# ── Auth ─────────────────────────────────────────────────
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9

# ── Utilities ────────────────────────────────────────────
tiktoken>=0.7.0
langdetect>=1.0.9

# ── Database Driver (sync PostgreSQL) ────────────────────
psycopg2-binary>=2.9.9
